Create aligningHighResMAP
Purpose: Given MAP image figure out where it is in the 10x MAP. Just for fun(And because I do not want to find where every single MAP is within the Larger scale by hand) Currently skeletonization works okay. Trying to fit the skeletons with a spline (Autodetected) so that splines can be aligned based on curvature. Some kind of version of ICP maybe. Also this could be used for Aygul to align her MAP onto her in vivo later on.