Convex Hull Version of Outline
Allows for outline that gets all of the area that is contained. Outlines work from center of pixel. Not what I want for in vivo but maybe I'll use this later for map. New method since this doesn't work will involve eroding the mask and then just looking at the difference.